+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.3:
+## New Content:
+- The Genie Warlock: Elemental Gift
+## New Features:
+- The item replacer can now check additional compendiums. 	
+  
 # 0.2.2:
 ## New Content:
 - Blight